+++ eggs/Naaya/eggs/Naaya/CHANGELOG.rst
@@ -1,5 +1,7 @@
 3.2.9 (unreleased)
 -------------------
+* Add `gdata` dependency because of broken objects [moregale]
+* Update script: remove `gdata` object instances [moregale]
 
 3.2.8 (2012-12-17)
 -------------------