+++ eggs/Naaya/eggs/Naaya/CHANGELOG.rst
@@ -1,4 +1,7 @@
-3.2.9 (unreleased)
+3.2.10 (unreleased)
+-------------------
+
+3.2.9 (2012-12-17)
 -------------------
 * Add `gdata` dependency because of broken objects [moregale]
 * Update script: remove `gdata` object instances [moregale]